Commit generated code to the repository
Replace the build script with an integration test that fails if the
generated code is out of date. This has several advantages:

* No need for downstream users to compile the build dependencies
* No need for downstream users to compile the build script
* No need for downstream users to run the code generation step
* The generated code is more usable in an IDE
* CI will fail if the generated code is out of date

The main downside is probably that the generated code is kept in
the repository, but this seems of relatively minimal impact,
especially if the code generator/source files don't change often.